+++ src/deflate/mod.rs
@@ -3,11 +3,15 @@ use libz_sys;
 use miniz_sys;
 use libc::c_int;
 use std::cmp::max;
+use zopfli;
+
+pub mod libz_stream;
+pub mod miniz_stream;
 
 /// Decompress a data stream using the DEFLATE algorithm
 pub fn inflate(data: &[u8]) -> Result<Vec<u8>, PngError> {
     let mut input = data.to_owned();
-    let mut stream = super::libz_stream::Stream::new_decompress();
+    let mut stream = libz_stream::Stream::new_decompress();
     let mut output = Vec::with_capacity(data.len());
     loop {
         match stream.decompress_vec(input.as_mut(), output.as_mut()) {
@@ -29,10 +33,8 @@ pub fn deflate(data: &[u8], zc: u8, zm: u8, zs: u8, zw: u8) -> Result<Vec<u8>, P
     let mut output = Vec::with_capacity(max(1024, data.len() / 20));
     if zs == 0 || zs == 1 {
         // Miniz performs 1-2 orders of magnitude better for strategies 0 and 1
-        let mut stream = super::miniz_stream::Stream::new_compress(zc as c_int,
-                                                                   zw as c_int,
-                                                                   zm as c_int,
-                                                                   zs as c_int);
+        let mut stream =
+            miniz_stream::Stream::new_compress(zc as c_int, zw as c_int, zm as c_int, zs as c_int);
         loop {
             match stream.compress_vec(input.as_mut(), output.as_mut()) {
                 miniz_sys::MZ_OK => output.reserve(max(1024, data.len() / 20)),
@@ -42,10 +44,8 @@ pub fn deflate(data: &[u8], zc: u8, zm: u8, zs: u8, zw: u8) -> Result<Vec<u8>, P
         }
     } else {
         // libz performs an order of magnitude better for strategies 2 and 3
-        let mut stream = super::libz_stream::Stream::new_compress(zc as c_int,
-                                                                  zw as c_int,
-                                                                  zm as c_int,
-                                                                  zs as c_int);
+        let mut stream =
+            libz_stream::Stream::new_compress(zc as c_int, zw as c_int, zm as c_int, zs as c_int);
         loop {
             match stream.compress_vec(input.as_mut(), output.as_mut()) {
                 libz_sys::Z_OK => output.reserve(max(1024, data.len() / 20)),
@@ -58,4 +58,24 @@ pub fn deflate(data: &[u8], zc: u8, zm: u8, zs: u8, zw: u8) -> Result<Vec<u8>, P
     output.shrink_to_fit();
 
     Ok(output)
+}
+
+pub fn zopfli_deflate(data: &[u8]) -> Result<Vec<u8>, PngError> {
+    let mut output = Vec::with_capacity(max(1024, data.len() / 20));
+    let options = zopfli::Options::default();
+    match zopfli::compress(&options, &zopfli::Format::Zlib, data, &mut output) {
+        Ok(_) => (),
+        Err(_) => return Err(PngError::new("Failed to compress in zopfli")),
+    };
+    output.shrink_to_fit();
+    Ok(output)
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+/// DEFLATE algorithms supported by oxipng
+pub enum Deflaters {
+    /// Use the Zlib DEFLATE implementation
+    Zlib,
+    /// Use the better but slower Zopfli implementation
+    Zopfli,
 }