+++ CHANGELOG.md
@@ -6,6 +6,8 @@
  - [SEMVER_MAJOR] Bump minimum rustc version to 1.9.0, required by dependencies
  - [SEMVER_MINOR] Allow calling optimization presets via crate using `Options::from_preset`
  - [SEMVER_MAJOR] Return proper `PngError` type which implements `std::error::Error` from `Result`s
+ - [SEMVER_MAJOR] Rename module `deflate::deflate` to `deflate`
+ - [SEMVER_MINOR] Add zopfli option (`-Z`)
  - Performance optimizations
 
 **Version 0.10.0**